Update toolchain and base-layout to use /usr
Mere started out without using /usr for much, on the basis that /usr
is a historical artifact that has no particular meaning. The experiment
worked well, but going forward, it appears that /usr is still expected
and comfortable for most people, there are a few workarounds that we
have to do to ignore it (technically it still existed anyway) and the
benefits of doing so are minimal at best.

Avoid unnecessarily suprising new users by keeping /usr as the main
installation prefix.