Fix bug in build.sh
realpath_cd was printing out the directory that was changed to when
calling `cd -` in addition to the output from `pwd -P`.

Also adding -x to the top of the script for gathering info about the
pipeline environment. The docker builds need to be done with at least a
fake root account so that permissions in the generated packages are
correct. So this begins and investigation into the correct solution
there.