Simplify build scripts by removing uid/gid config
Previously, uid and gid were added so that files created in mounted
volumes would retain the host system ownership for easier
management/deletion of those files. A side effect, however is that
created packages would contain files with incorrect ownership.

Remove the added complexity and require that the containers would be run
with the default uid/gid, which should be root.