Refactored the gen_changelog script
Adjust the script to produce a new changelog format. It doesn't seem
important to list a maintainer email in each entry. Instead, the main
entry should be primarily tied to a release version and the date can be
included for reference. The new format will look like:

```txt
1.2.3-1 (2021-08-29)

    This is a log message indicating a change.
    This is another change.
    Changes could also be listed as:
    - Change 1
    - Change 2
```

The script will first reformat an existing ChangeLog, and then attempt
to prompt for and add a new entry if one doesn't exist for the package version
and release combination.