[s6] Adjust mdevd configuration
Remove the mdevd service file and configuration. Those will be added to
the mdevd package in a future commit. This follows a similar pattern as
the utmps package and allows mdevd to be optionally installed, in case a
user wishes to run without kernel modules.

Also adjust the call to mdevd-coldplug to wait until mdevd has signaled
that it is up